Separate template helpers into their own module
